Don't try to kill server if we don't have a valid pid.

+++ noaweather/weatherServer.py
@@ -144,11 +144,13 @@ if __name__ == "__main__":
         server = SocketServer.UDPServer(("localhost", conf.server_port), clientHandler)
     except socket.error:
         print "Can't bind address: %s, port: %d." % ("localhost", conf.server_port)
-        print 'Killing old server with pid %d' % conf.weatherServerPid
-        os.kill(conf.weatherServerPid, signal.SIGTERM)
-        time.sleep(2)
-        conf.serverLoad()
-        server = SocketServer.UDPServer(("localhost", conf.server_port), clientHandler)
+        
+        if conf.weatherServerPid is not False:
+            print 'Killing old server with pid %d' % conf.weatherServerPid
+            os.kill(conf.weatherServerPid, signal.SIGTERM)
+            time.sleep(2)
+            conf.serverLoad()
+            server = SocketServer.UDPServer(("localhost", conf.server_port), clientHandler)
     
     # Save pid
     conf.weatherServerPid = os.getpid()
@@ -170,7 +172,4 @@ if __name__ == "__main__":
     conf.serverSave()
     print 'Server stoped.'
     logfile.close()
-    
-    
-
